// crypto-tracker-web-client/src/model/global-metrics-table.ts
export class GlobalMetricsTable {
  asset: string;
  symbol: string;
  percentageOfGlobalCap: any;
  current: number;
  tangibleCurrency: number;
  worldsBillionaires: number;
  gold: number;
  stocks: number;
  narrowMoney: number;
  broadMoney: number;
  realEstate: number;
  wealth: number;
  derivatives: number;
}